Generate the definitions separately
Initially it was the authors understanding that the definitions for the
various protobuf libraries were fairly consistent. However, there are a
number of issues in the golang repository that indicate that this is not
the case:

  - https://github.com/golang/protobuf/issues/646
  - https://github.com/golang/protobuf/issues/391

These bugs indicate a more general set of behaviours that need to be
customized when generating the libraries.

This commit implements such custom generation, fixing the bug mentioned
in the golang #646 issue, as well as some other nicities with the
swagger specification