Generate the API gateway golang library
Part of the ongoing requirements for this API is to allow both
application/json and gRPC interfaces, while keeping the work to a
minimal to support both those interfaces.

To do this, we use the gRPC api gateway that transforms application/json
data into gRPC requests. This allows expressing protobuf as a swagger
definition, and consuming the otherwise gRPC only API via
swagger.

== Design Notes ==

=== Included here ===

It arguably makes sense to generate and store the gRPC gateway somewhere
else. However, practical limitations on generating the code mean it's
simply easy to vendor the gateway generated in the same repository.