Move the gRPC gateway into the go library
The upstream protobuf library allows generating a library that
transcodes application/json into gRPC.

However, it expects to be part of the same package as the other
definitions; specifically as part of the service definition. The types
are generated this way, and other references hint at this problem.

Rather than attempt to delimit this manually, this commit instead
chooses to simply package the libraries together as the gRPC gateway
expects.

This is apparently the most common use case for this tooling.